moved login and signup to the api routes, added dynamic greeting, authenticated user
