+++ controllers/api/index.js
@@ -7,6 +7,10 @@ const task = require('./task');
 const role = require('./role');
 const employeeTask = require('./employeeTask');
 
+//! api login 
+const login = require('./login');
+const signup = require('./signup');
+
 
 // api routes
 router.use('/employee', employee);
@@ -14,6 +18,10 @@ router.use('/department', department);
 router.use('/task', task);
 router.use('/role', role);
 router.use('/employeeTask', employeeTask);
+
+// ! api login/registration routes
+router.use('/login', login);
+router.use('/signup', signup);
 
 
 module.exports = router;
