+++ controllers/htmlRoutes/index.js
@@ -2,14 +2,17 @@
 const router = require('express').Router();
 
 // require the routes available in this folder 
-const login = require('./login');
-const signup = require('./signup');
+
+const signup = require('../api/signup');
 const home = require('./home'); 
+const loginPage = require('./loginPage'); 
 
 // html routes
-router.use('/', login);
+
+router.use('/', loginPage);
 router.use('/signup', signup);
 router.use('/home', home);
+
 
 
 module.exports = router;
