+++ public/assets/js/employeePage.js
@@ -1,0 +1,8 @@
+$(document).ready(function() {
+    // This file just does a GET request to figure out which user is logged in
+    // and updates the HTML on the page
+    $.get("/api/login/user_data").then(function(data) {
+      $(".employee-name").text(data.email);
+    });
+  });
+
