+++ views/home.handlebars
@@ -59,8 +59,8 @@
                         <div class="col-12 ">
                             <div
                                 class="jumbotron jumbotron-fluid pb-4 pt-2 px-4 font-weight-bold  border border-primary rounded-lg">
-                                <h1 class=" text-white">Welcome name!! <span class=" text-white" id="currentDay">
-                                        display the current time here</span></h1>
+                                {{!-- employee-name will be here  --}}
+                                <h1 class=" text-white">Welcome, <span class="employee-name"></span>!!</h1>
                                 <h1 class="display-4"><span class="city">this just displays basic info for the user to
                                         see</span> (<span class="day0">8/07/20</span>)</h1>
                                 <div>this will show a picture of the user if we have time </div>
@@ -82,10 +82,12 @@
             <div class="col-12">
                 <h1 class="ml-n3">Other tasks that need to be done!</h1>
                 <div class="row d-flex justify-content-between text-center">
-                    <div class="col-sm-12 mb-2 col-md-6 border border-info bg-primary font-weight-bold  rounded-lg p-2 pb-3">
-                    one task
+                    <div
+                        class="col-sm-12 mb-2 col-md-6 border border-info bg-primary font-weight-bold  rounded-lg p-2 pb-3">
+                        one task
                     </div>
                 </div>
             </div>
         </div>
     </div>
+    <script src="/assets/js/employeePage.js"></script>
